requirements-lock.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==23.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
